+++ tests/windows_aggregator_resize_2_test.py
